+++ fluent_pages/tests/__init__.py
@@ -7,3 +7,4 @@ from .urldispatcher import UrlDispatcherTests, UrlDispatcherNonRootTests
 from .menu import MenuTests
 from .modeldata import ModelDataTests
 from .plugins import PluginTests, PluginUrlTests
+from .templatetags import TemplateTagTests