+++ fluent_pages/tests/testapp/urls.py
@@ -2,4 +2,5 @@ from django.conf.urls.defaults import patterns, url, include
 
 urlpatterns = patterns('',
     url(r'', include('fluent_pages.urls')),
+    url(r'^404/$', 'django.views.defaults.page_not_found'),
 )