+++ fluent_pages/__init__.py
@@ -1,4 +1,5 @@
-VERSION = (0, 1, 0)
+# following PEP 386
+__version__ = "0.8.3"
 
 # Do some sane version checking
 import django