+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changes in version 0.8.3 (dev)
 
 * Fix circular imports for some setups that import ``fluent_pages.urlresolvers`` early.
 * Fix initial south migrations, added missing dependencies.
+* Fix using ``{% render_menu %}`` at 404 pages.
 
 
 Changes in version 0.8.2